Move container test fixtures to root conftest
Consolidates pytest_addoption, pytest_report_header, and related fixtures
in tests/conftest.py. This fixes the unrecognized --container-image error
(options must be in root conftest) and avoids mypy duplicate module issue.
